Re-implement chat message reading/writing
